# Optimize dBFS/amplitude conversion functions