`std::pow(10.f, x)` and `log10f` are painfully slow functions when run per-sample, as is commonly required via `dbfsToAmp` and `ampToDbfs`.  This PR sacrifices a completely negligible amount of precision for a significant performance boost for several LMMS plugins:

```
static inline float fastPow10f(float x)
{
	return expf(2.302585092994046f * x);
}

static inline float fastLog10f(float x)
{
	return logf(x) * 0.4342944819032518f;
}
```

Notice that these are mathematically equivalent operations, not approximations.  The lost precision, which is very nearly 0, is solely due to the limited precision of floats.  

-----
Compiled via GCC with -O2:

```
1 billion iterations of log10f(x):
31,138,340,305      cycles
8.419502684 seconds time elapsed

1 billion iterations of fastLog10f(x):
15,626,908,607      cycles
4.226729945 seconds time elapsed
```

fastLog10f is **1.993x** as fast on this hardware.

```
1 billion cycles of pow(10.f, x):
26,823,368,193      cycles
7.258407044 seconds time elapsed

1 billion cycles of fastPow10f(x):
16,135,296,386      cycles
4.315204418 seconds time elapsed
```

fastPow10f is **1.662x** as fast on this hardware.
(This might even provide a larger benefit on older hardware, I've seen performance benefits of up to 3x reported by others.)

-----

It might be worth exploring some faster and less-precise alternatives in the future, but what this PR provides is a significant improvement for no actual downsides whatsoever.